AAstat {seqinr}	R Documentation

To Get Some Protein Statistics

Description

Returns simple protein sequence information including the number of residues, the percentage physico-chemical classes and the theoretical isoelectric point. The functions ignore ambiguous amino acids (e.g. "B", "Z", "X", "J").

Usage

AAstat(seq, plot = TRUE)

Arguments

seq
a protein sequence as a vector of upper-case chars

plot
if TRUE, plots the presence of residues splited by physico-chemical classes along the sequence.

Value

A list with the three following components:

Compo
A factor giving the amino acid counts.

Prop
A list giving the percentage of each physico-chemical classes (Tiny, Small, Aliphatic, Aromatic, Non-polar, Polar, Charged, Positive, Negative).

Pi
The theoretical isoelectric point

Author(s)

D. Charif, J.R. Lobry

References

citation("seqinr")

See Also

computePI, SEQINR.UTIL, SeqFastaAA

Examples

  seqAA <- read.fasta(file = system.file("sequences/seqAA.fasta", package = "seqinr"),
   seqtype = "AA")
  AAstat(seqAA[[1]])

[Package seqinr version 4.2-36 Index]