JLO {seqinr}	R Documentation

Forensic Genetic Profile Raw Data

Description

This is an example of raw data for a human STR genetic profile at 16 loci (viz. D8S1179, D21S11, D7S820, CSF1PO, D3S1358, TH01, D13S317, D16S539, D2S1338, D19S433, vWA, TPOX, D18S51, Amelogenin, D5S818, FGA) which are commonly used in forensic sciences for individual identifications.

Usage

data(JLO)

Format

A list with 3 components.

Header

a list corresponding to the header in the ABIF file

Directory

a data.frame corresponding to the Directory in the ABIF file

Data

a list with all raw data in the ABIF file.

Details

This dataset is the expected result when reading the file 2_FAC321_0000205983_B02_004.fsa with the function read.abif. This dataset is used for the quality check of this function.

Author(s)

J.R. Lobry

Source

The DNA source is from the author so that there are no privacy concern. Data were kindly provided by the INPS (Institut National de Police Scientifique) which is the national forensic sciences institute in France. Experiments were done at the LPS (Laboratoire de Police Scientifique de Lyon) in 2008.

References

citation("seqinr")

Anonymous (2006) Applied Biosystem Genetic Analysis Data File Format. Available at https://www.thermofisher.com/de/de/home/brands/applied-biosystems.html. Last visited on 03-NOV-2008.

See Also

function read.abif to import files in ABIF format, data gs500liz for internal size standards, data ECH for the corresponding allelic ladder, data identifiler for allele names in the allelic ladder.

Examples

data(JLO)

[Package seqinr version 4.2-36 Index]