SeqAcnucWeb {seqinr}	R Documentation

Sequence coming from a remote ACNUC data base

Description

as.SeqAcnucWeb is called by many functions, for instance by query, and should not be directly called by the user. It creates an object of class SeqAcnucWeb. is.SeqAcnucWeb returns TRUE if the object is of class SeqAcnucWeb.

Usage

as.SeqAcnucWeb(object, length, frame, ncbigc)
is.SeqAcnucWeb(object)

Arguments

object
a string giving the name of a sequence present in the data base

length
a string giving the length of the sequence present in the data base

frame
a string giving the frame of the sequence present in the data base

ncbigc
a string giving the ncbi genetic code of the sequence present in the data base

Value

as.SeqAcnucWeb returns an object sequence of class SeqAcnucWeb. Note that as from seqinR 1.1-3 the slot socket has been deleted to save space for long lists.

Author(s)

D. Charif, J.R. Lobry

References

citation("seqinr")

Examples

  ## Not run: # Need internet connection
  choosebank("emblTP")
  mylist <- query("mylist", "sp=felis catus et t=cds et o=mitochondrion")
  stopifnot(is.SeqAcnucWeb(mylist$req[[1]]))
  closebank()

## End(Not run)

[Package seqinr version 4.2-36 Index]