SeqFastaAA {seqinr}	R Documentation

AA sequence in Fasta Format

Description

as.SeqFastaAA is called by the function as read.fasta. It creates an object of class SeqFastaAA. is.SeqFastaAA returns TRUE if the object is of class SeqFastaAA. summary.SeqFastaAA gives the AA composition of an object of class SeqFastaAA.

Usage

as.SeqFastaAA(object, name = NULL, Annot = NULL)
is.SeqFastaAA(object)
## S3 method for class 'SeqFastaAA'
summary(object,...)

Arguments

object
a vector of chars representing a biological sequence

name
NULL a character string specifying a name for the sequence

Annot
NULL a character string specifying some annotations for the sequence

...
additional arguments affecting the summary produced

Value

as.SeqFastaAA returns an object sequence of class SeqFastaAA. summary.SeqFastaAA returns a list which the following components:

composition
the AA counting of the sequence

AA.Property
the percentage of each group of amino acid in the sequence. By example, the groups are small, tiny, aliphatic, aromatic ...

Author(s)

D. Charif

References

citation("seqinr")

Examples

  s <- read.fasta(file = system.file("sequences/seqAA.fasta", package = "seqinr"), seqtype="AA")
  is.SeqFastaAA(s[[1]])
  summary(s[[1]])
  myseq <- s2c("MSPTAYRRGSPAFLV*")
  as.SeqFastaAA(myseq, name = "myseq", Annot = "blablabla")
  myseq

[Package seqinr version 4.2-36 Index]