SeqFrag {seqinr}	R Documentation

Class for sub-sequences

Description

as.SeqFrag is called by all methods of getFrag, but not directly by the users. It creates an object sequence of class SeqFrag.

Usage

 as.SeqFrag(object, begin, end, name)
 is.SeqFrag(object)

Arguments

object
an object sequence of class seqFastadna, seqFastaAA, seqAcnucWeb or seqFrag

begin
the first base of the fragment to get

end
the last base of the fragment to get

name
the name of the sequence

Value

as.SeqFrag returns a biological sequence with the following attributes:

seqMother
the name of the sequence from which the sequence comes

begin
the position of the first base of the fragment on the mother sequence

end
the position of the last base of the fragment on the mother sequence

class
SeqFrag which is the classfor sub-sequence

is.SeqFrag returns TRUE if the object is of class Seqfrag.

Author(s)

D. Charif, J.R. Lobry

References

citation("seqinr")

See Also

getFrag, getLength, getName, getSequence, getTrans

Examples

 s <- read.fasta(file = system.file("sequences/malM.fasta", package = "seqinr"))
 getFrag(s[[1]], 1, 10) 	

[Package seqinr version 4.2-36 Index]