alllistranks {seqinr}	R Documentation

To get the count of existing lists and all their ranks on server

Description

This is a low level function to get the total number of list and all their ranks in an opened database.

Usage

alllistranks(socket = autosocket(), verbose = FALSE)
alr(socket = autosocket(), verbose = FALSE)

Arguments

socket
an object of class sockconn connecting to a remote ACNUC database (default is a socket to the last opened database).

verbose
if TRUE, verbose mode is on

Details

This low level function is usually not used directly by the user.

Value

A list with two components:

count
count of existing lists

rank
their rank

Author(s)

J.R. Lobry

References

citation("seqinr")

See Also

choosebank, query

Examples

 ## Not run: # Need internet connection
 choosebank("emblTP")
 tmp1 <- query("tmp1", "sp=Borrelia burgdorferi", virtual = TRUE)
 tmp2 <- query("tmp2", "sp=Borrelia burgdorferi", virtual = TRUE)
 tmp3 <- query("tmp3", "sp=Borrelia burgdorferi", virtual = TRUE)
 (result <- alllistranks())
 stopifnot(result$count == 3)   # Three ACNUC lists
 stopifnot(result$ranks == 2:4) # Starting at rank 2
 #
 # Summay of current lists defined on the ACNUC server:
 #
 sapply(result$ranks, getliststate)
 closebank()
 
## End(Not run)

[Package seqinr version 4.2-36 Index]