autosocket {seqinr}	R Documentation

Returns a socket to the last opened database

Description

This is a low level function that is mainly used to select automatically the last opened ACNUC database for functions using sockets.

Usage

autosocket()

Value

An object of class sockconn.

Author(s)

J.R. Lobry

References

https://doua.prabi.fr/databases/acnuc.html

citation("seqinr")

See Also

choosebank, connections.

Examples

## Not run:  #Need internet connection
  choosebank("emblTP")
  autosocket()
  closebank()
  
## End(Not run)

[Package seqinr version 4.2-36 Index]