baselineabif {seqinr}	R Documentation

Estimation of baseline value

Description

This function tries to estimate the baseline value for RFU data from capillary electrophoresis whith the heuristic that the most common value is the baseline.

Usage

baselineabif(rfu, maxrfu = 1000)

Arguments

rfu
a numeric vector of signal value

maxrfu
signal values greater or equal to maxrfu are forced to NA

Value

A single numeric value for the estimated baseline.

Author(s)

J.R. Lobry

See Also

JLO for a dataset example, plotabif to plot this kind of data, peakabif to estimate peak parameters.

Examples

data(JLO)
rfu <- JLO$Data$DATA.1
bl <- baselineabif(rfu)
plot(1:length(rfu), rfu, type = "l", 
  xlab = "Time [datapoint units]",
  ylab = "Signal [RFU]",
  main = "Example of baseline estimates")
abline(h = bl, col="red", lty = 2)
legend("topright", inset = 0.02, "Baseline estimate", lty = 2, col = "red")

[Package seqinr version 4.2-36 Index]