closebank {seqinr}	R Documentation

To close a remote ACNUC database

Description

This function tries to close a remote ACNUC database.

Usage

closebank(socket = autosocket(), verbose = FALSE)

Arguments

socket
an object of class sockconn connecting to a remote ACNUC database (default is a socket to the last opened database).

verbose
Logical. If TRUE, verbose mode is on

Author(s)

J.R. Lobry

References

citation("seqinr")

See Also

choosebank

Examples

  ## Not run: # Need internet connection
   choosebank("emblTP")
   closebank()
  
## End(Not run)

[Package seqinr version 4.2-36 Index]