countfreelists {seqinr}	R Documentation

The number of free lists available and annotation lines in an ACNUC server

Description

Returns the number of free lists available list of names of annotation lines in the opened ACNUC database.

Usage

countfreelists(socket = autosocket())
cfl(socket = autosocket())

Arguments

socket
an object of class sockconn connecting to a remote ACNUC database (default is a socket to the last opened database).

Value

a list with the following 2 components:

free
numeric. The number of free lists

annotlines
vector of strings. Names of annotation lines

Author(s)

J.R. Lobry

References

https://doua.prabi.fr/databases/acnuc.html

citation("seqinr")

See Also

choosebank, query

Examples

## Not run:  # Need internet connection
  choosebank("emblTP")
  (rescountfreelists <- countfreelists())
  stopifnot(all(rescountfreelists$annotlines ==
   c("ALL", "AC",  "PR",  "DT",  "KW",  "OS",  "OC",
   "OG",  "RN",  "RC",  "RP",  "RX", "RG",  "RA",  "RT",  "RL",  "DR",
   "CC",  "AH",  "AS",  "FH",  "FT",  "CO",  "SQ", "SEQ")))
  closebank()
  
## End(Not run)

[Package seqinr version 4.2-36 Index]