countsubseqs {seqinr}	R Documentation

Number of subsequences in an ACNUC list

Description

Returns the number of subsequences in the ACNUC list of rank lrank.

Usage

countsubseqs(lrank, socket = autosocket())
css(lrank, socket = autosocket())

Arguments

lrank
the rank of the ACNUC list to consider.

socket
an object of class sockconn connecting to a remote ACNUC database (default is a socket to the last opened database).

Value

Numeric.

Author(s)

J.R. Lobry

References

https://doua.prabi.fr/databases/acnuc.html

citation("seqinr")

See Also

choosebank, query, glr to get a list rank from its name.

Examples

## Not run:  # Need internet connection
  choosebank("emblTP")
  mylist<-query("mylist", "N=@", virtual = TRUE) # select all (seqs + subseqs)
  mylist$nelem   # 14138094 seqs + subseqs
  stopifnot(mylist$nelem == 14138094)
  css(glr("mylist")) # 1604500 subsequences only
  stopifnot(css(glr("mylist")) == 1604500)
  closebank()
  
## End(Not run)

[Package seqinr version 4.2-36 Index]