draw.oriloc {seqinr}	R Documentation

Graphical representation for nucleotide skews in prokaryotic chromosomes.

Description

Graphical representation for nucleotide skews in prokaryotic chromosomes.

Usage

draw.oriloc(ori, main = "Title",
  xlab = "Map position in Kb",
  ylab = "Cumulated combined skew in Kb", las = 1, las.right = 3,
  ta.mtext = "Cumul. T-A skew", ta.col = "pink", ta.lwd = 1,
  cg.mtext = "Cumul. C-G skew", cg.col = "lightblue", cg.lwd = 1,
  cds.mtext = "Cumul. CDS skew", cds.col = "lightgreen", cds.lwd = 1,
  sk.col = "black", sk.lwd = 2,
  add.grid = TRUE, ...)

Arguments

ori
A data frame obtained with the oriloc function.

main
The main title of the plot.

xlab
The x-axis title.

ylab
The y-axis title.

las
The style of axis labels for the bottom and left axes.

las.right
The style of axis labels for the right axis.

ta.mtext
The marginal legend for the TA skew.

ta.col
The color for the TA skew.

ta.lwd
The line width for the TA skew.

cg.mtext
The marginal legend for the CG skew.

cg.col
The color for the CG skew.

cg.lwd
The line width for the CG skew.

cds.mtext
The marginal legend for the CDS skew.

cds.col
The color for the CDS skew.

cds.lwd
The line width for the CDS skew.

sk.col
The color for the cumulated combined skew.

sk.lwd
The line width for the cumulated combined skew.

add.grid
Logical, if TRUE a vertical grid is added to the plot.

...
Further arguments are passed to the function plot.

Author(s)

J.R. Lobry

References

citation("seqinr")

See Also

oriloc, rearranged.oriloc, extract.breakpoints

Examples

## Not run:  # need internet connection
#
# Example with Chlamydia trachomatis complete genome
#
  ori <- oriloc()
  draw.oriloc(ori)
#
# The same, using more options from function draw.oriloc()
#
draw.oriloc(ori, 
  main = expression(italic(Chlamydia~~trachomatis)~~complete~~genome),
  ta.mtext = "TA skew", ta.col = "red",
  cg.mtext = "CG skew", cg.col = "blue",
  cds.mtext = "CDS skew", cds.col = "seagreen",
  add.grid = FALSE)

## End(Not run)

[Package seqinr version 4.2-36 Index]