gbk2g2.euk {seqinr}	R Documentation

Conversion of a GenBank format file into a glimmer-like one. Eukaryotic version.

Description

This function reads a file in GenBank format and converts the features corresponding to CDS (Coding Sequences) into a format similar to glimmer program output. This function is specifically made for eukaryotic sequences, i.e. with introns.

Usage

gbk2g2.euk(gbkfile = system.file("sequences/ame1.gbk", package ="seqinr"),
g2.coord = "g2.coord")

Arguments

gbkfile
The name of the GenBank file

g2.coord
The name of the output file

Details

This function returns the coordinates of the exons annotated in the GenBank format file.

Value

A data frame with three columns will be written to the g2.coord file. The first column corresponds to the name of the gene, given in the GenBank file through the /gene feature. The second and third column contain the start and the stop position of the exon.

Author(s)

J.R. Lobry, A. Necşulea

References

citation("seqinr")

See Also

oriloc, gbk2g2

Examples

 ## Not run:  gbk2g2.euk() 

[Package seqinr version 4.2-36 Index]