gcT {seqinr}	R Documentation

GC content and temperature in bacteria

Description

This data set was used in Galtier and Lobry (1997) to study the relationship between the optimal growth temperature of bacteria and their G+C content at the genomic level and locally were selection is active to maintain secondary structures in the stems of RNAs.

Format

gcT is a list containing the 9 following components:

species

is a data frame containing the optimal growth temperature and genomic G+C content for 772 bacterial species. Detailled explanations for this table and the following are available in the README component.

genus

is a data frame containing the optimal growth temperature and genomic G+C content for 224 bacterial genus.

details

is a data frame with more information, see README.

gc16S

is a data frame containing the optimal growth temperature and stems G+C content for 16S RNA from 165 bacterial genus.

gctRNA

is a data frame containing the optimal growth temperature and stems G+C content for tRNA from 51 bacterial genus.

gc23S

is a data frame containing the optimal growth temperature and stems G+C content for 23S RNA from 38 bacterial genus.

gc5S

is a data frame containing the optimal growth temperature and stems G+C content for 5S RNA from 71 bacterial genus.

README

is the original README file from ftp://biom3.univ-lyon1.fr/pub/datasets/JME97/ last updated 13-MAY-2002.

importgcT

is the R script used to import data.

Source

Galtier, N. & Lobry, J.R. (1997). Relationships between genomic G+C content, RNA secondary structures, and optimal growth temperature in prokaryotes. Journal of Molecular Evolution 44:632-636.

Data imported into seqinr with the R script given in the last component of the dataset by J.R. Lobry on 09-OCT-2016.

References

citation("seqinr")

Examples

data(gcT)

[Package seqinr version 4.2-36 Index]