getKeyword {seqinr}	R Documentation

Generic function to get keywords associated to sequences

Description

Get keywords from an ACNUC server.

Usage

getKeyword(object, ...)
## S3 method for class 'SeqAcnucWeb'
getKeyword(object, ..., socket = autosocket())

Arguments

object
an object of the class SeqAcnucWeb, or a list of them, or the object resulting from query

socket
an object of class sockconn connecting to a remote ACNUC database (default is a socket to the last opened database by choosebank).

...
further arguments passed to or from other methods

Value

getKeyword returns a vector of strings containing the keyword(s) associated to a sequence.

Author(s)

D. Charif, J.R. Lobry, L. Palmeira

References

citation("seqinr")

See Also

SeqAcnucWeb

Examples

#
# List all available methods for getKeyword generic function:
#
   methods(getKeyword)
#
# Example of keyword extraction from an ACNUC server:
#
  ## Not run: 
  # Need internet connection
  choosebank("emblTP")
  fc<-query("fc", "sp=felis catus et t=cds et o=mitochondrion")
  getKeyword(fc$req[[1]])
  # Should be: 
  # [1] "DIVISION ORG" "RELEASE 62"   "CYTOCHROME B" "SOURCE"       "CDS"
  closebank()  

## End(Not run)

[Package seqinr version 4.2-36 Index]