getType {seqinr}	R Documentation

To get available subsequence types in an opened ACNUC database

Description

This function returns all subsequence types (e.g. CDS, TRNA) present in an opened ACNUC database, using default database if no socket is provided.

Usage

getType(socket = autosocket())

Arguments

socket
an object of class sockconn connecting to a remote ACNUC database (default is a socket to the last opened database).

Value

a list containing a short description for each subsequence type.

Author(s)

D. Charif, J.R. Lobry

References

citation("seqinr")

See Also

choosebank, query

Examples

## Not run: 
# Need internet connection
  choosebank("emblTP")
  getType()
## End(Not run)

[Package seqinr version 4.2-36 Index]