getlistrank {seqinr}	R Documentation

To get the rank of a list from its name

Description

This is a low level function to get the rank of a list on server from its name.

Usage

getlistrank(listname, socket = autosocket(), verbose = FALSE)
glr(listname, socket = autosocket(), verbose = FALSE)

Arguments

listname
the name of list on server

socket
an object of class sockconn connecting to a remote ACNUC database (default is a socket to the last opened database).

verbose
if TRUE, verbose mode is on

Details

This low level function is usually not used directly by the user.

Value

The rank of list named listname on server, or 0 if no list with this name exists.

Author(s)

J.R. Lobry

References

citation("seqinr")

See Also

choosebank, query

Examples

 ## Not run: 
 # Need internet connection
 choosebank("emblTP")
 MyListName <- query("MyListName", "sp=Borrelia burgdorferi", virtual = TRUE)
 (result <- getlistrank("MyListName"))
 stopifnot(result == 2)
 closebank()
 
## End(Not run)

[Package seqinr version 4.2-36 Index]