getliststate {seqinr}	R Documentation

Asks for information about an ACNUC list of specified rank

Description

Reply gives the type of list, its name, the number of elements it contains, and, for sequence lists, says whether the list contains only parent seqs (locus=T).

Usage

getliststate(lrank, socket = autosocket())
gls(lrank, socket = autosocket())
gln(lrank, ...)

Arguments

lrank
the name of the ACNUC list to modify

socket
an object of class sockconn connecting to a remote ACNUC database (default is a socket to the last opened database).

...
arguments passed to getliststate

Value

NA in case of problem and an warning is issued. When there is no problem a list with the following 4 components:

type
string. Type of ACNUC list (SQ, KW, SP)

name
string. ACNUC list name

count
numeric. Number of elements in ACNUC list

locus
logical. For ACNUC sequence lists TRUE means that the list contains only parent sequences. NA otherwise.

gln is a shortcut for getliststate(lrank, ...)$name

Author(s)

J.R. Lobry

References

https://doua.prabi.fr/databases/acnuc.html

citation("seqinr")

See Also

choosebank, query, alr, glr

Examples

## Not run: 
  ### Need internet connection
  choosebank("emblTP")
  mylist <- query("mylist", "sp=felis catus et t=cds", virtual=TRUE)
  getliststate(glr("mylist")) # SQ, MYLIST, 603, FALSE
  gln(glr("mylist")) # MYLIST (upper case letters on server)
  closebank()
  
## End(Not run)

[Package seqinr version 4.2-36 Index]