gfrag {seqinr}	R Documentation

Extract sequence identified by name or by number from an ACNUC server

Description

Get length characters from sequence identified by name or by number starting from position start (counted from 1).

Usage

gfrag(what, start, length, idby = c("name", "number"), socket = autosocket())

Arguments

what
A sequence name or number

start
Start position from 1

length
Number of requested characters (answer may be shorter)

idby
Is the sequence identified by name or number? Default to name

socket
an object of class sockconn connecting to a remote ACNUC database (default is a socket to the last opened database).

Value

A string of characters with at most length characters (may be shorter than asked for). NA is returned and a warning is issued in case of problem (non existent sequence for instance).

Author(s)

J.R. Lobry

References

https://doua.prabi.fr/databases/acnuc.html

citation("seqinr")

See Also

choosebank, query

Examples

## Not run: # Need internet connection
  choosebank("emblTP")
  gfrag("LMFLCHR36", start = 1, length = 3529852) -> myseq
  stopifnot(nchar(myseq) == 3529852)
  closebank()
  
## End(Not run)

[Package seqinr version 4.2-36 Index]