isenum {seqinr}	R Documentation

Get the ACNUC number of a sequence from its name or accession number

Description

Gives the ACNUC number of a sequence in the number element of the returned list. More informations are returned for subsequences corresponding to coding sequences.

Usage

isenum(what, idby = c("name", "access"), socket = autosocket())
isn(what, ...)
getNumber.socket(socket, name)
getAttributsocket(socket, name)

Arguments

what
a sequence name or a sequence accession number

idby
is the sequence identified by name or by accession number? Default to name

socket
an object of class sockconn connecting to a remote ACNUC database (default is a socket to the last opened database).

...
arguments passed to isenum.

name
a sequence name.

Value

A list whith the following 6 components:

number
numeric. The ACNUC number of the sequence.

length
numeric. The length of the sequence.

frame
numeric. The reading frame (0, 1, or 2) of the sequence for CDS.

gencode
numeric. ACNUC's genetic code (0 means universal) of the sequence for CDS.

ncbigc
numeric. NCBI's genetic code (0 means universal) of the sequence for CDS.

otheraccessmatches
logical. If TRUE it means that several sequences are attached to the given accession nunmber, and that only the ACNUC number of the first attached sequence is returned in the number component of the list.

isn(what, ...) is a shortcut for isenum(what, ...)$number.

As from seqinR 1.1-3 getNumber.socket and getAttributsocket are deprecated (a warning is issued).

Author(s)

J.R. Lobry

References

https://doua.prabi.fr/databases/acnuc.html

citation("seqinr")

See Also

choosebank, query

Examples

## Not run: 
  ### Need internet connection
  choosebank("emblTP")
  isenum("LMFLCHR36")
  isn("LMFLCHR36")
  stopifnot(isn("LMFLCHR36") == 13682678)
  # Example with CDS:
  isenum("AB004237")
  
## End(Not run)

[Package seqinr version 4.2-36 Index]