kaksTorture {seqinr}	R Documentation

Expected numeric results for Ka and Ks in extreme cases

Description

This data set is what should be obtained when runing kaks() on the test file kaks-torture.fasta in the sequences directory of the seqinR package.

Usage

data(kaksTorture)

Format

A list with 4 components of class dist.

ka

Ka

ks

Ks

vka

variance for Ka

vks

variance for Ks

Source

See comments in kaks-torture.fasta for R code used to produce it.

References

citation("seqinr")

Examples

data(kaksTorture)
kaks.torture <- read.alignment(file = system.file("sequences/kaks-torture.fasta", 
  package = "seqinr"), format = "fasta")
#
# Failed on windows :
#
# stopifnot(identical(kaksTorture, kaks(kaks.torture)))
# stopifnot(identical(kaksTorture, kaks(kaks.torture, rmgap = FALSE)))

[Package seqinr version 4.2-36 Index]