knowndbs {seqinr}	R Documentation

Description of databases known by an ACNUC server

Description

Returns, for each database known by the server, its name (a valid value for the bank argument of choosebank), availability (off means temporarily unavailable), and description.

Usage

knowndbs(tag = c(NA, "TP", "TEST", "DEV"), socket = autosocket())
kdb(tag = c(NA, "TP", "TEST", "DEV"), socket = autosocket())

Arguments

tag
default to NA, see details

socket
an object of class sockconn connecting to a remote ACNUC database (default is a socket to the last opened database).

Details

When the optional tag argument is used, only databases tagged with the given string are listed; when this argument is NA (by default), only untagged databases are listed. The tag argument thus allows to identify series of special purpose (tagged) databases, in addition to default (untagged) ones.

Value

A dataframe with 3 columns:

bank
string. Valid bank values known by the ACNUC server

status
string. "on" means available, "off" means temporarily unavailable

info
string. short description of the database

Author(s)

J.R. Lobry

References

https://doua.prabi.fr/databases/acnuc.html

citation("seqinr")

The full list of untagged and tagged databases is here : https://doua.prabi.fr/databases/acnuc/banques_raa.php.

See Also

choosebank when called without arguments.

Examples

## Not run: 
  ### Need internet connection
  choosebank("emblTP")
  kdb()
  closebank()
  
## End(Not run)

[Package seqinr version 4.2-36 Index]