peakabif {seqinr}	R Documentation

Extraction of Peak locations, Heights and Surfaces from ABIF data

Description

Simple peak location for data imported with the read.abif function using cubic spline interpolation.

Usage

peakabif(abifdata, 
  chanel, 
  npeak, 
  thres = 400/yscale, 
  fig = TRUE,
  chanel.names = c(1:4,105),
  DATA = paste("DATA", chanel.names[chanel], sep = "."),
  tmin = 1/tscale,
  tmax = abifdata$Data[["SCAN.1"]]/tscale,
  tscale = 1000, 
  yscale = 1000,
  irange = (tmin*tscale):(tmax*tscale),
  y = abifdata$Data[[DATA]][irange]/yscale,
  method = "monoH.FC",
  maxrfu = 1000,
  ...)

Arguments

abifdata
the result returned by read.abif

chanel
the dye number

npeak
the expected number of peaks

thres
scaled threshold value

fig
logical: should localized peaks be plotted

chanel.names
numbers extensions used for the DATA

DATA
names of the DATA components

tmin
scaled starting time for the time axis

tmax
scaled ending time for the time axis

tscale
scale factor for the time axis

yscale
scale factor for the y-axis (RFU)

irange
indices of data to be plotted

y
values used for the y-axis

method
method to be used by splinefun

maxrfu
argument passed to baselineabif

...
arguments forwarded to plot

Value

Returns invisibly a list with the unscaled values for the locations of peaks, heights of peaks and surfaces of peaks and baseline estimate. The peak location are in datapoint units, that is an integer starting at 1 for the first experimental point, 2 for the second experimental point, etc. However, due to interpolation between points the estimated peak location is usually not an integer.

Author(s)

J.R. Lobry

References

citation("seqinr")

See Also

function read.abif to import files in ABIF format, plotabif to plot them, data gs500liz for internal size standards, data identifiler for allele names in the allelic ladder, data JLO for an example of an individual sample file, data ECH for an example of an allelic lader.

Examples

data(JLO)
JLO.maxis <- peakabif(JLO, 5, npeak = 14, tmin = 2.7, thres = 0.1)$maxis

[Package seqinr version 4.2-36 Index]