phylip {seqinr}	R Documentation

Example of results obtained after a call to read.alignment

Description

This data set gives an example of a amino acids alignment obtained after a call to the function read.alignment on an alignment file in "phylip" format.

Usage

data(phylip)

Format

A List of class alignment

Source

http://evolution.genetics.washington.edu/phylip.html

References

http://evolution.genetics.washington.edu/phylip.html

[Package seqinr version 4.2-36 Index]