plot.SeqAcnucWeb {seqinr}	R Documentation

To Plot Subsequences on the Parent Sequence

Description

This function plots all the type of subsequences on a parent sequence. Subsequences are represented by colored rectangle on the parent sequence. For example, types could be CDS, TRNA, RRNA .... In order to get all the types that are available for the selected database, use getType.

Usage

## S3 method for class 'SeqAcnucWeb'
plot(x, types = getType()$sname, socket = autosocket(), ...)

Arguments

x
A sequence of class SeqAcnucWeb

types
The type of subsequences to plot. Default value is to consider all possible subsequence types.

socket
an object of class sockconn connecting to a remote ACNUC database (default is a socket to the last opened database).

...
not currently used

Value

An invisible list giving, for each subsequence, its position on the parent sequence.

Author(s)

D. Charif, J.R. Lobry

References

https://doua.prabi.fr/databases/acnuc.html

citation("seqinr")

See Also

getType, query

Examples

## Not run: 
  ### Need internet connection
  choosebank("emblTP")
  mylist <- query("mylist", "AC=AB078009")
  plot(mylist$req[[1]])
  
## End(Not run)

[Package seqinr version 4.2-36 Index]