prepgetannots {seqinr}	R Documentation

Select annotation lines in an ACNUC database

Description

This function is called before using getAnnot or modifylist with a scan type operation to select the annotation lines to be returned or scanned.

Usage

prepgetannots(what = "all", setfor = c("scan", "getannots"),
                       socket = autosocket(), verbose = FALSE)
pga(what = "all", setfor = c("scan", "getannots"),
                       socket = autosocket(), verbose = FALSE)

Arguments

what
the default "all" means that all annotation lines are selected. This can be more specific, see details.

setfor
this is used when what has its default "all" value. The behaviour is different for getAnnot and modifylist with a scan type operation: annotations but not sequences are scanned, but sequences can be returned by getAnnot. The default value is "scan".

socket
an object of class sockconn connecting to an ACNUC server

verbose
logical, if TRUE mode verbose is on

Details

The names of annotation lines in the opened ACNUC database is returned by countfreelists, they are forced to upper case letters by prepgetannots when supplied with the what argument.

For the EMBL/SWISSPROT format, keys are: ALL, AC, DT, KW, OS, OC, OG, OH, RN, RC, RP, RX, RA, RG, RT, RL, DR, AH, AS, CC, FH, FT, SQ, SEQ.

For GenBank: ALL, ACCESSION, VERSION, KEYWORDS, SOURCE, ORGANISM, REFERENCE, AUTHORS, CONSRTM, TITLE, JOURNAL, PUBMED, REMARK, COMMENT, FEATURES, ORIGIN, SEQUENCE.

For FT (embl, swissprot) and FEATURES (GenBank), one or more specific feature keys can be specified using lines with only uppercase and such as

FEATURES|CDS FT|TRNA

Keys ALL and SEQ/SEQUENCE stand for all annotation and sequence lines, respectively. For the scan operation, key ALL stand for the DE/DEFINITION lines, and SEQ/SEQUENCE cannot be used (annotations but not sequence are scanned).

Value

The function returns invisibly the annotation lines names.

Author(s)

J.R. Lobry

References

citation("seqinr")

See Also

getAnnot, modifylist, countfreelists

Examples

 ## Not run: # Need internet connection
  choosebank("genbank")
  mylist <- query("mylist","n=AQF16SRRN")
  pga() # We want to scan all annotations, including FEATURES
  mylist <- modifylist("mylist", operation = "strain", type = "scan")
  mylist$nelem # should be 1
 
## End(Not run)

[Package seqinr version 4.2-36 Index]