prettyseq {seqinr}	R Documentation

Text representation of a sequence from an ACNUC server

Description

To get a text representation of sequence of rank num and of its subsequences, with bpl bases per line (default = 60), and with optional translation of protein-coding subsequences

Usage

prettyseq(num, bpl = 60, translate = TRUE, socket = autosocket())

Arguments

num
rank of the sequence in the ACNUC database

bpl
number of base per line

translate
should coding sequences be translated?

socket
an object of class sockconn connecting to a remote ACNUC database (default is a socket to the last opened database).

Value

An invisible vector of string. The output is redirected to the console.

Author(s)

J.R. Lobry

References

https://doua.prabi.fr/databases/acnuc.html

citation("seqinr")

See Also

choosebank, query

Examples

## Not run: 
  ### Need internet connection
  choosebank("emblTP")
  prettyseq(111)
  
## End(Not run)

[Package seqinr version 4.2-36 Index]