readsmj {seqinr}	R Documentation

Low level function to read ACNUC SMJYT index files

Description

Extract informations from the SMJYT index file for status, molecule, journal, year, type, organelle, division, and db structure information.

Usage

readsmj(socket = autosocket(), num = 2, nl = 10, recnum.add = FALSE, nature.add = TRUE,
plong.add = FALSE, libel.add = FALSE, sname.add = FALSE, all.add = FALSE)

Arguments

socket
an object of class sockconn connecting to a remote ACNUC database (default is a socket to the last opened database).

num
rank number of first record.

nl
number of records to read.

recnum.add
to extract record numbers.

nature.add
to extract as a factor with human understandable levels the nature of the name. Unordered levels are: status, molecule, journal, year, type, organelle, division and dbstrucinfo.

plong.add
to extract the plong.

libel.add
to extract the label of the name.

sname.add
to extract the short version of the name, that is without the first two characters.

all.add
to extract all (all flags set to TRUE).

Value

A data.frame with requested columns.

Author(s)

J.R. Lobry

References

See ACNUC physical structure at: https://doua.prabi.fr/databases/acnuc/structure.html.

citation("seqinr")

See Also

choosebank to start a session and readfirstrec to get the total number of records.

[Package seqinr version 4.2-36 Index]