s2c {seqinr}	R Documentation

conversion of a string into a vector of chars

Description

This is a simple utility function to convert a single string such as "BigBang" into a vector of chars such as c("B", "i", "g", "B", "a", "n", "g").

Usage

s2c(string)

Arguments

string
a string of chars

Value

a vector of chars. If supplied argument is not a single string, a warning is issued and NA returned.

Author(s)

J.R. Lobry

References

citation("seqinr")

See Also

c2s

Examples

stopifnot(all(s2c("BigBang") == c("B", "i", "g", "B", "a", "n", "g")))

[Package seqinr version 4.2-36 Index]