savelist {seqinr}	R Documentation

Save sequence names or accession numbers into a file

Description

This function retrieves all sequence names or all accession number from an ACNUC list and saves them into a file.

Usage

savelist(lrank, type = c("N", "A"),
                     filename = paste(gln(lrank), ifelse(type == "N", "mne", "acc"),
		     sep = "."),socket = autosocket(), warnme = TRUE)

Arguments

lrank
the rank of the ACNUC list to consider.

type
use "N" for sequence names (mnemonics) and "A" for accession numbers. Default is "N".

filename
a string of character giving the name of the file to save results.

socket
an object of class sockconn connecting to a remote ACNUC database (default is a socket to the last opened database).

warnme
if TRUE a message is issued on the console when complete.

Value

none.

Author(s)

J.R. Lobry

References

https://doua.prabi.fr/databases/acnuc.html

citation("seqinr")

See Also

choosebank, query, glr to get a list rank from its name, clfcd for the inverse operation of savelist

Examples

## Not run: 
  ### Need internet connection
  choosebank("emblTP")
  mylist <- query("mylist", "sp=felis catus et t=cds", virtual=TRUE)
  savelist(glr("mylist"))
  # 603 sequence mnemonics written into file: MYLIST.mne
  savelist(glr("mylist"), type = "A")
  # 603 sequence accession numbers written into file: MYLIST.acc
  
## End(Not run)

[Package seqinr version 4.2-36 Index]