setlistname {seqinr}	R Documentation

Sets the name of an ACNUC list identified by its rank

Description

This is a low level function to set the name of a list from an ACNUC server. It should not be used directly by end users.

Usage

setlistname(lrank, name = "list1", socket = autosocket())

Arguments

lrank
the list rank on the ACNUC server

name
the name to use for this list

socket
an object of class sockconn connecting to a remote ACNUC database (default is a socket to the last opened database).

Value

A single numeric value corresponding to:

NA
Empty answer from server.

0
OK.

3
if another list with that name already existed and was deleted.

4
no list of rank lrank exists.

Author(s)

J.R. Lobry

References

https://doua.prabi.fr/databases/acnuc.html

citation("seqinr")

See Also

choosebank, query, glr

Examples

## Not run: 
  ### Need internet connection
  choosebank("emblTP")
  mylist <- query("mylist", "sp=felis catus et t=CDS", virtual = TRUE)
  # Change list name on server:
  setlistname(lrank = glr("mylist"), name = "feliscatus") # 0, OK.
  glr("mylist") # 0, list doesn't exist no more.
  glr("feliscatus") # 2, this list exists.
  # Note the danger here: the object mylist is still present in the user workspace
  # while the corresponding list was deleted from server.
  
## End(Not run)

[Package seqinr version 4.2-36 Index]