stresc {seqinr}	R Documentation

Utility function to escape LaTeX special characters present in a string

Description

This function returns a vector of strings in which LaTeX special characters are escaped, this was useful in conjunction with xtable.

Usage

stresc(strings)

Arguments

strings
A vector of strings to deal with.

Value

Returns a vector of strings with escaped characters within each string.

Author(s)

J.R. Lobry

References

citation("seqinr")

See Also

s2c

Examples

  stresc("MISC_RNA")
  stresc(c("BB_0001","BB_0002"))

[Package seqinr version 4.2-36 Index]