synsequence {seqinr}	R Documentation

Random synonymous coding sequence generation

Description

Generates a random synonymous coding sequence, according to a certain codon usage bias

Usage

synsequence(sequence, numcode = 1, ucoweight = NULL)

Arguments

sequence
A nucleic acids sequence

numcode
The genetic code number as in translate

ucoweight
A list of weights containing the desired codon usage bias as generated by ucoweight

Value

a sequence translating to the same protein sequence as the original one (cf. translate), but containing synonymous codons

Author(s)

L. Palmeira

References

citation("seqinr")

See Also

ucoweight

Examples

  data(ec999)
  sequence=ec999[1][[1]]
  synsequence(sequence,1,ucoweight(sequence))

[Package seqinr version 4.2-36 Index]