tablecode {seqinr}	R Documentation

to plot genetic code as in textbooks

Description

This function plots a genetic code table as in textbooks, that is following the order T > C > A > G so that synonymous codons are almost always in the same boxes.

Usage

tablecode(numcode = 1, urn.rna = s2c("TCAG"), dia = FALSE, latexfile = NULL,
label = latexfile, size = "normalsize", caption = NULL,
preaa = rep("", 64), postaa = rep("", 64), 
precodon = preaa, postcodon = postaa)

Arguments

numcode
The genetic code number as in translate

urn.rna
The letters to display codons, use s2c("UCAG") if you want the code in terms of RNA sequence

latexfile
The name of a LaTex file if you want to redirect the output

label
The label for the LaTeX table

size
The LaTex size of characters for the LaTeX table

preaa
A string to insert before the amino-acid in the LaTeX table

postaa
A string to insert after the amino-acid in the LaTeX table

precodon
A string to insert before the codon in the LaTeX table

postcodon
A string to insert after the codon in the LaTeX table

caption
The caption of the LaTeX table

dia
to produce a yellow/blue plot for slides

Details

The codon order for preaa, postaa, precodon, and postcodon should be the same as in paste(paste(rep(s2c("tcag"), each =16), s2c("tcag"), sep = ""), rep(s2c("tcag"), each = 4), sep = "")

Author(s)

J.R. Lobry

References

citation("seqinr")

See Also

translate, syncodons

Examples

#
# Show me the standard genetic code:
#

 tablecode()

[Package seqinr version 4.2-36 Index]