toycodon {seqinr}	R Documentation

A toy example of codon counts in three coding sequences

Description

This is a toy data set to illustrate synonymous and non-synonymous codon usage analyses.

Usage

data(toyaa)

Format

A data frame with 3 observations (coding sequences) for 10 codons.

Source

Created for release 1.0-4 of seqinr's vignette.

References

citation("seqinr")

Examples

data(toycodon)

[Package seqinr version 4.2-36 Index]