ucoweight {seqinr}	R Documentation

Weight of each synonymous codon

Description

Returns a list containing, for each of the 20 amino acids + STOP codon, the codon usage bias of each of the synonymous codon according to a given codon sequence.

Usage

ucoweight(sequence, numcode = 1)

Arguments

sequence
A nucleic acids sequence

numcode
The genetic code number as in translate

Value

a list containing, for each of the 20 amino acids and STOP codon (list tags), the weight of each synonymous codon (including the original one).

Author(s)

L. Palmeira

References

citation("seqinr")

See Also

synsequence

Examples

  data(ec999)
  ucoweight(ec999[1][[1]])

[Package seqinr version 4.2-36 Index]