where.is.this.acc {seqinr}	R Documentation

Scans databases for a given sequence accession number

Description

This function loops over all availabale ACNUC databases to look for a given sequence accession number. This is useful when you have a sequence accession number and you don't know in which database it is present.

Usage

where.is.this.acc(acc,  stopAtFirst = TRUE, ...)

Arguments

acc
An accession number as a string of characters such as "NC_001416".

stopAtFirst
Logical. If TRUE, the default, the function stops at the first database where the accession number is found.

...
Arguments passed to the function choosebank.

Value

The function resturns invisibly a vector of strings of characters for the names of the ACNUC databases in which the accession number was found.

Author(s)

J.R. Lobry

References

citation("seqinr")

See Also

choosebank to open a given ACNUC database.

Examples

  ## Not run: # Need internet connection
  where.is.this.acc("NC_001416") # first found in phever2dna bank (2016-06-01)
  
## End(Not run)

[Package seqinr version 4.2-36 Index]